SX Design: resolve "system" theme internally
The purpose of this change is to resolve "system" theme into "light" or "dark" for the underlying components. System theme is only for the user to define that the SX Design should follow system settings. However, all the components care only about "light" or "dark" (not system). This was causing issues in several components (`Skeleton` for example).

adeira-source-id: dd6fc8926a3ba6c85dab1c3221eed82cd567f78c