SX Design: forward React refs in `ButtonLink`
I noticed that Next.js links were not working very well for `ButtonLink` and `Link` components. This change adds a `forwardRef` to `ButtonLink` and `onClick` callback to both of the components which fixes the issue (for more details see: https://nextjs.org/docs/api-reference/next/link#if-the-child-is-a-function-component).

Rest of the changes are just minor tweaks.

adeira-source-id: 33e4709fb7252c81c6b3a5154c795d0a30301c6a