SX Design: add new component `LayoutGrid` with a default spacing
I first version of a layout component creating a traditional CSS grid.
This is a very common pattern (at least in my apps), so I would like to
extract it and make it available amongst other layout components.

Ideas for the future development is to make it more customizable and
improve responsibility when it comes to mobiles and tablets (basically
change grid and spacing sizes).

adeira-source-id: 3798228aff5d57e614f665ea9b2ace597e49a130