SX Design: improve `Tooltip` positioning
Previously, we didn't take into account the children width which caused
weird positioning for wider elements. This changes fixes is by correctly
positioning the tooltip in the middle of the children.

There are some other minor tweaks. Most notably, it now takes into
account client scroll position (client bounding context is not enough)
and it can position the tooltip when it doesn't fit from the right.

adeira-source-id: adb6f0605a99987b0a2994534f5bf566d0f78194