SX Design: improve `ProductCard` styles
First, seems like I broke the styles a little when I was removing the
deprecated `Heading` component. This change fixes it.

Second, the card looked a bit weird when the product names are long.
So I tried to improve it a little by introducing variable title
size and limiting the width of the highlight.

adeira-source-id: dd71d3b33ff1aff08d2da6f22253568b0e3cf6c0