SX Design: remove `useAccessibleColors` hook from `Text` component
I simply do not know how to implement it properly and there are several
reported issues around it. So I'd like to remove it from SX Design and
leave this for the users (at least for now).

It's still being used inside `ColorShowcase` but that's just for our
Storybook, not the user-facing code.

adeira-source-id: 4ae331ffca6fd5eec108c3ce06df2227ab539f25