SX Design: unify h1-h6 styles
This was bothering me for a while: `Text` component was using default
browser styles when using h1-h6. That means that you had to manually
adjust the sizes/weights to match rest of the system and in some cases
it was not even achievable.

This change unifies the styles. So for example: `as=p, size=48` looks
like `h1` and vice-versa.

adeira-source-id: c3ab65fc7b827f739cb3698ed9096b37cee284f9